Abgabedateien/script/requirements.txt
mysql-connector-python==8.0.32
paho-mqtt==1.6.1
